Решение домашнего задания. Урок 5

+++ Lessons-2/MetricsAgent/Models/CpuMetric.cs
@@ -4,5 +4,5 @@ public class CpuMetric
 {
     public int Id { get; set; }
     public int Value { get; set; }
-    public int Time { get; set; }
+    public long Time { get; set; }
 }
